+++ framework-crates/objc2-io-kit/src/lib.rs
@@ -54,15 +54,15 @@ pub(crate) type FourCharCode = u32;
 
 // device/device_types.h
 #[allow(dead_code, non_camel_case_types)]
-pub(crate) type io_name_t = *mut [core::ffi::c_char; 128];
+pub(crate) type io_name_t = [core::ffi::c_char; 128];
 #[allow(dead_code, non_camel_case_types)]
-pub(crate) type io_string_t = *mut [core::ffi::c_char; 512];
+pub(crate) type io_string_t = [core::ffi::c_char; 512];
 #[allow(dead_code, non_camel_case_types)]
-pub(crate) type io_struct_inband_t = *mut [core::ffi::c_char; 4096];
+pub(crate) type io_struct_inband_t = [core::ffi::c_char; 4096];
 
 // uuid/uuid_t.h
 #[allow(dead_code, non_camel_case_types)]
-pub(crate) type uuid_t = [u8; 16]; // Usage sites are all in structs
+pub(crate) type uuid_t = [u8; 16];
 
 /// [Apple's documentation](https://developer.apple.com/documentation/iokit/io_object_null?language=objc)
 #[cfg(feature = "libc")]